requirements.txt
google-genai
opencv-python #open cv- computer vision
pyaudio
pillow
mss
streamlit
streamlit-webrtc
python-dotenv